requirements.txt
crewai
litellm
python-dotenv
streamlit
pandas